Changed index to main folder so it loads properly, changed front end main.jsx and app.jsx so they initialize React, updated register and login forms in front end and server.js in backend to run on 5173 local host
