+++ client/src/App.jsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import HomePage from './pages/HomePage/HomePage.jsx';
 import DashboardPage from './pages/Dashboard/DashboardPage.jsx';
