--- client/src/components/Login/LoginForm.css
@@ -1,1 +1,0 @@
-
