--- client/src/components/Register/RegisterForm.css
@@ -1,1 +1,0 @@
-
