--- client/src/components/SearchBar/SearchBar.css
@@ -1,1 +1,0 @@
-
