+++ client/src/pages/HomePage/HomePage.jsx
@@ -1,8 +1,7 @@
 import React from 'react';
 import LoginForm from '../../components/Login/LoginForm.jsx';
 import RegisterForm from '../../components/Register/RegisterForm.jsx';
-import './HomePage.css';
-
+import '../../styles/styles.css';
 
 function HomePage() {
   return (
