+++ package.json
@@ -4,34 +4,33 @@
     "description": "A full-stack web app with React, Express, PostgreSQL, and Sequelize",
     "main": "server.js",
     "scripts": {
-      "start": "concurrently \"npm run server\" \"npm run client\"",
-      "server": "cd server && node server.js",
-      "client": "cd client && npm run dev",
-      "install-client": "cd client && npm install",
-      "install-server": "cd server && npm install",
-      "install": "npm run install-client && npm run install-server"
+        "start": "concurrently \"npm run server\" \"npm run client\"",
+        "server": "cd server && node server.js",
+        "client": "cd client && npm run dev",
+        "install-client": "cd client && npm install",
+        "install-server": "cd server && npm install",
+        "install": "npm run install-client && npm run install-server"
     },
     "dependencies": {
-      "axios": "^0.27.2",
-      "concurrently": "^7.0.0",
-      "cors": "^2.8.5",
-      "dotenv": "^16.0.0",
-      "express": "^4.18.2",
-      "pg": "^8.7.3",
-      "pg-hstore": "^2.3.4",
-      "sequelize": "^6.22.0",
-      "jsonwebtoken": "^8.5.1",
-      "bcryptjs": "^2.4.3",
-      "sequelize-cli": "^6.3.0"
+        "bcryptjs": "^2.4.3",
+        "concurrently": "^7.0.0",
+        "cors": "^2.8.5",
+        "dotenv": "^16.0.0",
+        "express": "^4.18.2",
+        "jsonwebtoken": "^9.0.2",
+        "pg": "^8.7.3",
+        "pg-hstore": "^2.3.4",
+        "sequelize": "^6.22.0",
+        "sequelize-cli": "^6.3.0"
     },
     "devDependencies": {
-      "vite": "^4.0.0",
-      "react": "^18.0.0",
-      "react-dom": "^18.0.0",
-      "react-router-dom": "^6.0.0",
-      "axios": "^0.27.2"
+        "axios": "^1.7.9",
+        "react": "^18.0.0",
+        "react-dom": "^18.0.0",
+        "react-router-dom": "^6.0.0",
+        "vite": "^4.0.0"
     },
     "author": "megatron",
     "license": "ISC",
     "private": true
-  }
+}
