+++ server/config/config.js
@@ -1,6 +1,13 @@
 const { Sequelize } = require('sequelize');
 require('dotenv').config();
-
-const sequelize = new Sequelize(process.env.DB_URL);
-
+const sequelize = new Sequelize(process.env.DB_URL, {
+  logging: console.log, // Enable logging for debugging
+});
+sequelize.authenticate()
+  .then(() => {
+    console.log('Connection has been established successfully.');
+  })
+  .catch(err => {
+    console.error('Unable to connect to the database:', err);
+  });
 module.exports = sequelize;
