--- server/index.js
@@ -1,14 +1,0 @@
-// const express = require('express');
-
-// const app = express();
-
-// // Middleware to add user property to request
-// app.use((req, res, next) => {
-//     req.user = null; // Initialize user property
-//     next();
-// });
-
-// // Your routes and other middleware here
-
-// module.exports = app;
-
